requirements.txt
# Let's fix the requirements.txt file
@"
# Core ML and Deep Learning
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.1.0
pillow>=10.1.0

# Web Framework and API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1

# Streamlit (Alternative UI)
streamlit>=1.28.0
plotly>=5.17.0

# Data Processing and Augmentation
opencv-python-headless>=4.8.0
albumentations>=1.3.1
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and Monitoring
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Development and Testing
pytest>=7.4.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
"@ | Out-File -FilePath "requirements_fixed.txt" -Encoding UTF8

Write-Host "✅ Created fixed requirements file without sqlite3" -ForegroundColor Green
